\name{exsic.database}
\alias{exsic.database}
\title{Get hyperlink to database of passport data}
\description{Creates a hyperlink in markdown format.}
\usage{exsic.database()}

\value{A string containing the hyperlink to the file in markdown format.}

\author{Reinhard Simon}





