\name{index.collections}
\alias{index.collections}
\title{Creates the index of collectors and their specimes.}
\description{Creates the index of collectors and their specimes.}
\usage{index.collections()}

\value{A string containing the collectors formatted 
for the template file in markdown markup.}

\author{Reinhard Simon}
\note{Should only be used within the template file.}




