\name{index.countries}
\alias{index.countries}
\title{Creates a simple index of countries.}
\description{Creates a simple index of countries as text string.}
\usage{index.countries()}

\value{A string containing the species formatted 
for the template file in markdown markup.}

\author{Reinhard Simon}
\note{Should only be used within the template file.}




