\name{index.species}
\alias{index.species}
\title{Creates a simple index of species.}
\description{Creates an index of species as text string.}
\usage{index.species()}

\value{A string containing the species formatted 
for the template file in markdown markup.}

\author{Reinhard Simon}
\note{Should only be used within the template file.}




