\name{index.species.short}
\alias{index.species.short}
\title{A condensed list of species (listed within a line).}
\description{A condensed list of species (listed within a line).}
\usage{index.species.short()}

\value{A string containing the species formatted 
for the template file in markdown markup.}

\author{Reinhard Simon}
\note{Should only be used within the template file.}




