\name{index.specimens}
\alias{index.specimens}
\title{The list of specimens }
\description{List of specimens by species and country detailing collector,
collector number, date and location.}
\usage{index.specimens()}

\value{A string containing the specimen details formatted 
for the template file in markdown markup.}

\author{Reinhard Simon}
\note{Should only be used within the template file.}




